fix(ci): 修复格式检查与跨平台静态检查
按 Actions 失败日志格式化旧版真后端 E2E 用例，恢复 Prettier 门禁。

移除 Unix 密钥文件类型判断的冗余转换，使 Linux 静态检查与 Darwin 编译约束同时通过。

+++ apps/server/internal/secret/key_file_unix.go
@@ -116,7 +116,7 @@ func validatePrivateUnixHandle(fd int, directory bool) error {
 	if directory {
 		wantType = unix.S_IFDIR
 	}
-	if uint32(stat.Mode)&uint32(unix.S_IFMT) != wantType {
+	if uint32(stat.Mode)&unix.S_IFMT != wantType {
 		return fmt.Errorf("不是受支持的文件类型")
 	}
 	uid := os.Geteuid()
